Add files via upload
Pixel position of every individual source. S2 can be used as a reference source. With recent orbital elements, the position of Sgr A* can be estimated. From this, the absolute positions can be estimated. However, the relative positions can be used to determine the distance between the source of interest and Sgr A*. The distance is independent of using relative or absolute positions.